// src/enums/selectors.ts
'use strict';

/*
 * nemaaz
 *
 * (c) Sofiane Akbly <sofiane.akbly@digi-service.fr>
 *
 * Created by WebStorm on 27/09/2023 at 10:31
 * File /selectors
 */

export enum IshaSelector {
	ANGLE = 'ANGLE',
	MINUTES_AFTER_MAGHRIB = 'MINUTES_AFTER_MAGHRIB',
}

export enum MaghribSelector {
	ANGLE = 'ANGLE',
	MINUTES_AFTER_SUNSET = 'MINUTES_AFTER_SUNSET',
}